Fix combo command error breaking the whole script
I was told to add a strange enter command, 'combonew 1 1', which causes an error and seems to prevent CMD-X from ever loading again. I had to delete my configuration file as a result. This commit fixes that command and prevents other command errors from being abused this way in future.